[cosim] rename c/h files to not conflict with dromajo sources
